+++ CHANGES.rst
@@ -14,6 +14,8 @@ Changelog
 - added `remove-alpha` filter for removing any alpha channel from base images
 - the `sub-dir` reader now sets the current input
 - `from-instance-png-is` and `from-indexed-png-is` now ensure that the annotation is an image with an indexed palette
+- added `idc-layer-segments` tool to generate layer-segments output from image segmentation input files that require
+  processing in terms of file name alignment and base image augmentation
 
 
 0.1.0 (2025-10-31)